[Store][Azure] Rename `Azure` to `AzureSearch` in namespace
