+++ src/store/src/Bridge/AzureSearch/SearchStore.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Store\Bridge\Azure;
+namespace Symfony\AI\Store\Bridge\AzureSearch;
 
 use Symfony\AI\Platform\Vector\NullVector;
 use Symfony\AI\Platform\Vector\Vector;
